Extract add operation lifecycle into OperationRunner
