+++ tests/test_add_operation_lifecycle.py
@@ -1,20 +1,67 @@
 """Regression tests for the internal add-operation lifecycle phases."""
 
 from pathlib import Path
+from typing import Any
 
 import pytest
 
+import ogcat.catalog as catalog_module
 from ogcat import (
     ArtifactLocator,
     Catalog,
+    CatalogRecord,
     CatalogSpec,
     MetadataFieldDescription,
     OperationContext,
     OperationSource,
+    OperationState,
     PluginRegistry,
     RecordSchema,
     ValidationReport,
 )
+
+
+def test_run_add_operation_delegates_to_operation_runner(
+    tmp_path: Path,
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    """Catalog add setup delegates the lifecycle to OperationRunner."""
+    captured: dict[str, Any] = {}
+
+    class FakeRunner:
+        def __init__(self, *, dependencies: object, request: object) -> None:
+            captured["dependencies"] = dependencies
+            captured["request"] = request
+
+        def run(self) -> CatalogRecord:
+            request = captured["request"]
+            return CatalogRecord(
+                catalog="artifacts",
+                time_added="2026-05-14T00:00:00+00:00",
+                id="runner",
+                record_type=request.record_type,
+                locator=ArtifactLocator(kind="uri", value="s3://bucket/delegated.zarr"),
+            )
+
+    monkeypatch.setattr(catalog_module, "OperationRunner", FakeRunner)
+    catalog = Catalog.create(tmp_path / "catalog", CatalogSpec(catalog_name="artifacts"))
+
+    record = catalog.add_artifact(
+        record_type="external_reference",
+        locator=ArtifactLocator(kind="uri", value="s3://bucket/data.zarr"),
+        metadata={"title": "Delegated"},
+    )
+
+    dependencies = captured["dependencies"]
+    request = captured["request"]
+    assert dependencies.catalog_root == catalog.root
+    assert dependencies.hook_manager is catalog.hook_manager
+    assert request.transaction.repository is catalog.repository
+    assert request.commit is True
+    assert request.operation_type == "add_artifact"
+    assert request.record_type == "external_reference"
+    assert request.metadata == {"title": "Delegated"}
+    assert record.id == "runner"
 
 
 def test_add_file_lifecycle_preserves_hook_order_and_file_storage(tmp_path: Path) -> None:
@@ -163,6 +210,60 @@ def test_hook_failure_after_writer_rolls_back_artifact_and_record(tmp_path: Path
     assert rollback_calls == ["writer"]
     assert not target.exists()
     assert catalog.repository.all() == []
+
+
+def test_operation_runner_preserves_failure_audit_phase_and_rollback(tmp_path: Path) -> None:
+    """Runner-owned failures keep the existing failure phase and rollback audit."""
+
+    class FailingHook:
+        def before_record_write(self, context: OperationContext) -> None:
+            raise RuntimeError("stop before record")
+
+    source = tmp_path / "failure.nc"
+    source.write_text("payload", encoding="utf-8")
+    catalog = Catalog.create(
+        tmp_path / "catalog",
+        CatalogSpec(catalog_name="files"),
+        plugins=PluginRegistry([FailingHook()]),
+    )
+
+    with pytest.raises(RuntimeError, match="stop before record"):
+        catalog.add_file(source)
+
+    events = catalog.audit_events()
+    failure = next(event for event in events if event.event_type == "failure")
+    rollback_events = [event for event in events if event.event_type == "rollback"]
+    assert failure.details["phase"] == "before_record_write"
+    assert [event.message for event in rollback_events] == ["Rollback started.", "Rollback completed."]
+    assert catalog.repository.all() == []
+    assert source.exists()
+
+
+def test_operation_runner_preserves_caller_owned_transaction_on_hook_failure(tmp_path: Path) -> None:
+    """Caller-owned transactions remain staged when post-record hooks fail."""
+
+    class FailingAfterWriteHook:
+        def after_record_write(self, context: OperationContext) -> None:
+            raise RuntimeError("external transaction hook failure")
+
+    catalog = Catalog.create(
+        tmp_path / "catalog",
+        CatalogSpec(catalog_name="artifacts"),
+        plugins=PluginRegistry([FailingAfterWriteHook()]),
+    )
+
+    with catalog.transaction() as transaction:
+        with pytest.raises(RuntimeError, match="external transaction hook failure"):
+            catalog.add_artifact(
+                record_type="external_reference",
+                locator=ArtifactLocator(kind="uri", value="s3://bucket/data.zarr"),
+                transaction=transaction,
+            )
+
+        assert transaction.state is OperationState.STAGED
+        assert len(catalog.repository.all()) == 1
+        transaction.rollback()
+        assert catalog.repository.all() == []
 
 
 def test_validation_failure_runs_after_validate_and_stops_before_write(tmp_path: Path) -> None:
